# inkwell_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 gin + MySQL + Redis + JWT 的社区后端接口文档。\n所有接口返回统一结构: {\"code\":1000,\"message\":\"success\",\"data\":{}}。业务失败时 HTTP 状态码仍然是 200(只有访问不存在的路由才返回 404), 需要根据 code 判断结果, code 取值见 controller/code.go。\n需要登录的接口要在请求头带上 Authorization: Bearer \u003caccessToken\u003e; access token 失效(1006)时可以调用 /api/v1/refresh_token 刷新后重放请求。\n所有 64 位 ID(用户/帖子/评论)在 JSON 中都使用字符串传输, 用数字传会丢精度。",
        "title": "Inkwell 社区 API",
        "contact": {},
        "version": "0.1.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/comment": {
            "get": {
                "description": "按 post_id 查询某个帖子的全部评论(按时间正序), 或者按 ids 批量查询评论, 游客可访问。\n两个参数都没传或者 ID 非法时返回 code 1001。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID, 和 ids 二选一",
                        "name": "post_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "评论ID列表, 可以重复传多个, 例如 ?ids=1\u0026ids=2",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评论列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ApiComment"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "给帖子发表评论, 需要登录。内容最长 2000 字。\n失败时 HTTP 状态码仍是 200, 通过响应体里的 code 判断: 参数错误 1001, 未登录或 token 失效 1006/1008, 服务繁忙 1005。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "发表评论",
                "parameters": [
                    {
                        "description": "评论参数, parent_id 是要回复的评论ID, 不回复传 0",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCommentForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评论成功, 返回新评论ID",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.CommentIDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/post": {
            "get": {
                "description": "从 Redis 榜单分页读取帖子, order=score 按热度排序, order=time 按发帖时间排序, 游客可访问。\norder 取值不合法时返回 code 1001。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "帖子榜单",
                "parameters": [
                    {
                        "enum": [
                            "score",
                            "time"
                        ],
                        "type": "string",
                        "default": "score",
                        "description": "排序方式",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码, 从1开始",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子列表, 每页20条",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PostListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建帖子, 需要登录。标题最长 128 字, 内容最长 8192 字。\n失败时 HTTP 状态码仍是 200, 通过响应体里的 code 判断: 参数错误 1001, 未登录或 token 失效 1006/1008, 服务繁忙 1005。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "发帖",
                "parameters": [
                    {
                        "description": "发帖参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发帖成功, 返回新帖子ID",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PostIDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "description": "查询帖子详情, 并带上 Redis 里的实时票数与评论数, 游客可访问。帖子不存在时返回 code 1009。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "帖子详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ApiPostDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/post2": {
            "get": {
                "description": "直接查询 MySQL 的帖子列表, 按发帖时间倒序, 游客可访问",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "帖子列表(MySQL)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码, 从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页条数, 最大50",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PostListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.MyCode": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParams",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeServerBusy",
                "CodeInvalidToken",
                "CodeInvalidAuthFormat",
                "CodeNotLogin",
                "CodeNotFound",
                "CodeTooManyRequests"
            ]
        },
        "controller.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.MyCode"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ApiComment": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "0"
                },
                "author_name": {
                    "type": "string"
                },
                "comment_id": {
                    "type": "string",
                    "example": "0"
                },
                "content": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string",
                    "example": "0"
                },
                "post_id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "models.ApiPostDetail": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "0"
                },
                "author_name": {
                    "type": "string"
                },
                "comment_num": {
                    "type": "integer"
                },
                "community_id": {
                    "type": "integer"
                },
                "community_name": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "post_id": {
                    "description": "雪花算法生成的 ID 是 64 位整数, 超过 JS Number 的安全范围(2^53-1),\n序列化成字符串返回, 避免前端精度丢失后拿错误的 ID 去请求详情。",
                    "type": "string",
                    "example": "0"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "vote_num": {
                    "description": "VoteNum 是 Redis 中保存的净票数(赞成 - 反对, 可能为负); CommentNum 是评论数",
                    "type": "integer"
                }
            }
        },
        "models.CommentIDResponse": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "description": "CommentID 是雪花算法生成的 64 位 ID, 用字符串返回避免前端精度丢失",
                    "type": "string",
                    "example": "3458764513820540928"
                }
            }
        },
...
        "models.CreateCommentForm": {
            "type": "object",
            "required": [
                "content",
                "post_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "写得很清楚, 学到了"
                },
                "parent_id": {
                    "description": "ParentID 回复的父评论 ID, 不回复时传 0",
                    "type": "string",
                    "example": "0"
                },
                "post_id": {
                    "description": "PostID 是雪花算法生成的 64 位 ID, 需要用字符串传递, 否则会丢精度",
                    "type": "string",
                    "example": "3458764513820540928"
                }
            }
        },
        "models.CreatePostForm": {
            "type": "object",
            "required": [
                "community_id",
                "content",
                "title"
            ],
            "properties": {
                "community_id": {
                    "description": "CommunityID 版块 ID, 数字和字符串都接受",
                    "type": "integer",
                    "example": 1
                },
                "content": {
                    "type": "string",
                    "example": "gin 是一个用 Go 编写的 HTTP web 框架"
                },
                "title": {
                    "type": "string",
                    "example": "gin 框架入门"
                }
            }
        },
...
        "models.PostIDResponse": {
            "type": "object",
            "properties": {
                "post_id": {
                    "description": "PostID 是雪花算法生成的 64 位 ID, 用字符串返回避免前端精度丢失",
                    "type": "string",
                    "example": "3458764513820540928"
                }
            }
        },
        "models.PostListItem": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "0"
                },
                "author_name": {
                    "type": "string"
                },
                "comment_num": {
                    "type": "integer"
                },
                "community_id": {
                    "type": "integer"
                },
                "community_name": {
                    "type": "string"
                },
                "create_time": {
                    "description": "发帖时间(unix 秒)",
                    "type": "integer"
                },
                "post_id": {
                    "type": "string",
                    "example": "0"
                },
                "score": {
                    "description": "榜单排序分, 只有榜单接口会返回",
                    "type": "number"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vote_num": {
                    "description": "VoteNum 净票数(赞成 - 反对), 可能为负",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "取值格式 \"Bearer \u003caccessToken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}